Fix `Marshal.dump(RBTree.new)`
I hit a bug with `Marshal.dump(SortedSet.new)` when `rbtree` is installed.

SortedSet uses RBTree when it's present:
https://github.com/ruby/ruby/blob/3a083985a471ca3d8429146f9f18dead6747c203/lib/set.rb#L710-L714

Because of that, any class that encapsulates a SortedSet is affected.

The bug happens because `rb_marshal_dump` doesn't accept `limit` as its second
argument:
https://github.com/ruby/ruby/blob/3a083985a471ca3d8429146f9f18dead6747c203/marshal.c#L2272

I'm no C Ruby API expert and I am not sure what `port` means but this change
does fix a failing test in this project.